[ cleanup ] move Elab monad to separate module
